+++ src/Program.jl
@@ -124,19 +124,6 @@ function status_dependency(m::Module = @__MODULE__; exit_when_fail = false, verb
     check_dependency(m; exit_when_fail = exit_when_fail, verbose = verbose)
 end
 
-function generate_run_uuid(p::Program, inputs::Dict{String}, outputs::Dict{String})
-    out_uuid = UUID4
-    in_names = sort!([arg.name for arg in p.arg_inputs])
-    out_names = sort!([arg.name for arg in p.arg_outputs])
-    for name in in_names
-        out_uuid = uuid5(out_uuid, string(name, ":", inputs[name]))
-    end
-    for name in out_names
-        out_uuid = uuid5(out_uuid, string(name, ":", outputs[name]))
-    end
-    return out_uuid
-end
-
 """
     arg_completion(args::Vector{Arg}, xxputs::Dict{String})
 
@@ -450,24 +437,3 @@ end
 parse_verbose(verbose::Bool) = verbose ? :all : :none
 parse_verbose(::Nothing) = :none
 parse_verbose(verbose::AbstractString) = parse_verbose(Symbol(verbose))
-
-"""
-    isinputnewer(inputs::Dict, run_id_file::AbstractString) -> Bool
-
-Check whether any existing file (not dir) paths of `AbstractString` or `AbstractPath` in `inputs` are newer than `run_id_file`.
-"""
-function isinputnewer(inputs::Dict, run_id_file::AbstractString)
-    input_time = 0.0
-    for (k,v) in inputs
-        if v isa AbstractString || v isa AbstractPath
-            if isfile(v)
-                modified_time = mtime(v)
-                if modified_time > input_time
-                    input_time = modified_time
-                end
-            end
-        end
-    end
-
-    input_time > mtime(run_id_file)
-end